+++ screens/Introduction.tsx
@@ -20,37 +20,12 @@ import { CognitoHostedUIIdentityProvider } from "@aws-amplify/auth";
 import { RootStackScreenProps } from "../types";
 
 const Introduction = ({ navigation }: RootStackScreenProps<"Introduction">) => {
-  const [user, setUser] = useState(null);
-  const [customState, setCustomState] = useState(null);
-
   const handleFacebookLogin = async () => {};
   const handleGoogleLogin = async () => {
     return Auth.federatedSignIn({
       provider: CognitoHostedUIIdentityProvider.Google,
-      customState: "customState",
     });
   };
-  useEffect(() => {
-    const unsubscribe = Hub.listen("auth", ({ payload: { event, data } }) => {
-      switch (event) {
-        case "signIn":
-          setUser(data);
-          break;
-        case "signOut":
-          setUser(null);
-          break;
-        case "customOAuthState":
-          setCustomState(data);
-      }
-    });
-
-    Auth.currentAuthenticatedUser()
-      .then((currentUser) => setUser(currentUser))
-      .catch(() => console.log("Not signed in"));
-
-    return unsubscribe;
-  }, []);
-
   return (
     <View style={styles.container}>
       <SafeAreaView style={styles.header}>
